+++ lib/controllers/proxy.ts
@@ -1,0 +1,417 @@
+"use strict";
+
+// var basicAuth = require('basic-auth');
+var express = require("express");
+var defaultRequest = require("request");
+var url = require("url");
+var bodyParser = require("body-parser");
+var rangeCheck = require("range_check");
+
+var DO_NOT_PROXY_REGEX = /^(?:Host|X-Forwarded-Host|Proxy-Connection|Connection|Keep-Alive|Transfer-Encoding|TE|Trailer|Proxy-Authorization|Proxy-Authenticate|Upgrade|Expires|pragma|Strict-Transport-Security)$/i;
+var PROTOCOL_REGEX = /^\w+:\//;
+var DURATION_REGEX = /^([\d.]+)(ms|s|m|h|d|w|y)$/;
+var DURATION_UNITS: any = {
+  ms: 1.0 / 1000,
+  s: 1.0,
+  m: 60.0,
+  h: 60.0 * 60.0,
+  d: 24.0 * 60.0 * 60.0,
+  w: 7.0 * 24.0 * 60.0 * 60.0,
+  y: 365 * 24.0 * 60.0 * 60.0
+};
+/** Age to override cache instructions with for proxied files */
+var DEFAULT_MAX_AGE_SECONDS = 1209600; // two weeks
+
+/**
+ * Creates an express middleware that proxies calls to '/proxy/http://example' to 'http://example', while forcing them
+ * to be cached by the browser and overrwriting CORS headers. A cache duration can be added with a URL like
+ * /proxy/_5m/http://example which causes 'Cache-Control: public,max-age=300' to be added to the response headers.
+ *
+ * @param {Object} options
+ * @param {Array[String]} options.proxyableDomains An array of domains to be proxied
+ * @param {boolean} options.proxyAllDomains A boolean indicating whether or not we should proxy ALL domains - overrides
+ *                      the configuration in options.proxyDomains
+ * @param {String} options.proxyAuth A map of domains to tokens that will be passed to those domains via basic auth
+ *                      when proxying through them.
+ * @param {String} options.upstreamProxy Url of a standard upstream proxy that will be used to retrieve data.
+ * @param {String} options.bypassUpstreamProxyHosts An object of hosts (as strings) to 'true' values.
+ * @param {String} options.proxyPostSizeLimit The maximum size of a POST request that the proxy will allow through,
+                        in bytes if no unit is specified, or some reasonable unit like 'kb' for kilobytes or 'mb' for megabytes.
+ *
+ * @returns {*} A middleware that can be used with express.
+ */
+module.exports = function(options: any) {
+  var request = options.request || defaultRequest; //overridable for tests
+  var proxyAllDomains = options.proxyAllDomains;
+  var proxyDomains = options.proxyableDomains || [];
+  var proxyAuth = options.proxyAuth || {};
+  var proxyPostSizeLimit = options.proxyPostSizeLimit || "102400";
+  var appendParamToQueryString = options.appendParamToQueryString || {};
+
+  // If you change this, also change the same list in serverconfig.json.example.
+  // This page is helpful: https://en.wikipedia.org/wiki/Reserved_IP_addresses
+  var blacklistedAddresses = options.blacklistedAddresses || [
+    // loopback addresses
+    "127.0.0.0/8",
+    "::1/128",
+    // link local addresses
+    "169.254.0.0/16",
+    "fe80::/10",
+    // private network addresses
+    "10.0.0.0/8",
+    "172.16.0.0/12",
+    "192.168.0.0/16",
+    "fc00::/7",
+    // other
+    "0.0.0.0/8",
+    "100.64.0.0/10",
+    "192.0.0.0/24",
+    "192.0.2.0/24",
+    "198.18.0.0/15",
+    "192.88.99.0/24",
+    "198.51.100.0/24",
+    "203.0.113.0/24",
+    "224.0.0.0/4",
+    "240.0.0.0/4",
+    "255.255.255.255/32",
+    "::/128",
+    "2001:db8::/32",
+    "ff00::/8"
+  ];
+
+  //Non CORS hosts and domains we proxy to
+  function proxyAllowedHost(host: string) {
+    // Exclude hosts that are really IP addresses and are in our blacklist.
+    if (rangeCheck.inRange(host, blacklistedAddresses)) {
+      return false;
+    }
+
+    if (proxyAllDomains) {
+      return true;
+    }
+
+    host = host.toLowerCase();
+    //check that host is from one of these domains
+    for (var i = 0; i < proxyDomains.length; i++) {
+      if (
+        host.indexOf(proxyDomains[i], host.length - proxyDomains[i].length) !==
+        -1
+      ) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  function doProxy(
+    req: any,
+    res: any,
+    next: any,
+    retryWithoutAuth: any,
+    callback: any
+  ) {
+    var remoteUrlString = req.url.substring(1);
+
+    if (!remoteUrlString || remoteUrlString.length === 0) {
+      return res.status(400).send("No url specified.");
+    }
+
+    // Does the proxy URL include a max age?
+    var maxAgeSeconds = DEFAULT_MAX_AGE_SECONDS;
+    if (remoteUrlString[0] === "_") {
+      var slashIndex = remoteUrlString.indexOf("/");
+      if (slashIndex < 0) {
+        return res.status(400).send("No url specified.");
+      }
+
+      var maxAgeString = remoteUrlString.substring(1, slashIndex);
+      remoteUrlString = remoteUrlString.substring(slashIndex + 1);
+
+      if (remoteUrlString.length === 0) {
+        return res.status(400).send("No url specified.");
+      }
+
+      // Interpret the max age as a duration in Varnish notation.
+      // https://www.varnish-cache.org/docs/trunk/reference/vcl.html#durations
+      var parsedMaxAge = DURATION_REGEX.exec(maxAgeString);
+      if (!parsedMaxAge || parsedMaxAge.length < 3) {
+        return res.status(400).send("Invalid duration.");
+      }
+
+      var value = parseFloat(parsedMaxAge[1]);
+      if (value !== value) {
+        return res.status(400).send("Invalid duration.");
+      }
+
+      var unitConversion = DURATION_UNITS[parsedMaxAge[2]];
+      if (!unitConversion) {
+        return res.status(400).send("Invalid duration unit " + parsedMaxAge[2]);
+      }
+
+      maxAgeSeconds = value * unitConversion;
+    }
+
+    // Add http:// if no protocol is specified.
+    var protocolMatch = PROTOCOL_REGEX.exec(remoteUrlString);
+    if (!protocolMatch || protocolMatch.length < 1) {
+      remoteUrlString = "http://" + remoteUrlString;
+    } else {
+      var matchedPart = protocolMatch[0];
+
+      // If the protocol portion of the URL only has a single slash after it, the extra slash was probably stripped off by someone
+      // along the way (NGINX will do this).  Add it back.
+      if (remoteUrlString[matchedPart.length] !== "/") {
+        remoteUrlString =
+          matchedPart + "/" + remoteUrlString.substring(matchedPart.length);
+      }
+    }
+
+    var remoteUrl = url.parse(remoteUrlString);
+
+    // Copy the query string
+    remoteUrl.search = url.parse(req.url).search;
+
+    if (appendParamToQueryString[remoteUrl.host]) {
+      const options = appendParamToQueryString[remoteUrl.host];
+
+      for (var i = 0; i < options.length; i++) {
+        const option = options[i];
+        const re = new RegExp(option.regexPattern, "g");
+        const params = option.params;
+        if (re.test(remoteUrl.href)) {
+          const paramsString = Object.keys(params)
+            .map(key => key + "=" + params[key])
+            .join("&");
+          if (remoteUrl.search === null) {
+            remoteUrl.search = paramsString;
+          } else {
+            var joiner = remoteUrl.search.indexOf("?") >= 0 ? "&" : "?";
+            remoteUrl.search += joiner + paramsString;
+          }
+        }
+      }
+    }
+
+    if (!remoteUrl.protocol) {
+      remoteUrl.protocol = "http:";
+    }
+
+    var proxy;
+    if (
+      options.upstreamProxy &&
+      !(options.bypassUpstreamProxyHosts || {})[remoteUrl.host]
+    ) {
+      proxy = options.upstreamProxy;
+    }
+
+    // Are we allowed to proxy for this host?
+    if (!proxyAllowedHost(remoteUrl.host)) {
+      res
+        .status(403)
+        .send("Host is not in list of allowed hosts: " + remoteUrl.host);
+      return;
+    }
+
+    // encoding : null means "body" passed to the callback will be raw bytes
+
+    var proxiedRequest: any;
+    req.on("close", function() {
+      if (proxiedRequest) {
+        proxiedRequest.abort();
+      }
+    });
+
+    var filteredReqHeaders = filterHeaders(req.headers);
+    if (filteredReqHeaders["x-forwarded-for"]) {
+      filteredReqHeaders["x-forwarded-for"] =
+        filteredReqHeaders["x-forwarded-for"] +
+        ", " +
+        req.connection.remoteAddress;
+    } else {
+      filteredReqHeaders["x-forwarded-for"] = req.connection.remoteAddress;
+    }
+
+    // Remove the Authorization header if we used it to authenticate the request to terriajs-server.
+    if (
+      options.basicAuthentication &&
+      options.basicAuthentication.username &&
+      options.basicAuthentication.password
+    ) {
+      delete filteredReqHeaders["authorization"];
+    }
+
+    if (!retryWithoutAuth) {
+      var authRequired = proxyAuth[remoteUrl.host];
+      if (authRequired) {
+        if (authRequired.authorization) {
+          // http basic auth.
+          if (!filteredReqHeaders["authorization"]) {
+            filteredReqHeaders["authorization"] = authRequired.authorization;
+          }
+        }
+        if (authRequired.headers) {
+          // a mechanism to pass arbitrary headers.
+          authRequired.headers.forEach(function(header: any) {
+            filteredReqHeaders[header.name] = header.value;
+          });
+        }
+      }
+    }
+
+    proxiedRequest = callback(
+      remoteUrl,
+      filteredReqHeaders,
+      proxy,
+      maxAgeSeconds
+    );
+  }
+
+  function buildReqHandler(httpVerb: string) {
+    function handler(req: any, res: any, next: any) {
+      return doProxy(req, res, next, req.retryWithoutAuth, function(
+        remoteUrl: any,
+        filteredRequestHeaders: any,
+        proxy: any,
+        maxAgeSeconds: any
+      ) {
+        try {
+          var proxiedRequest = request({
+            method: httpVerb,
+            url: url.format(remoteUrl),
+            headers: filteredRequestHeaders,
+            encoding: null,
+            proxy: proxy,
+            body: req.body,
+            followRedirect: function(response: any) {
+              var location = response.headers.location;
+              if (location && location.length > 0) {
+                var parsed = url.parse(location);
+                if (proxyAllowedHost(parsed.host)) {
+                  // redirect is ok
+                  return true;
+                }
+              }
+              // redirect is forbidden
+              return false;
+            }
+          })
+            .on("socket", function(socket: any) {
+              socket.once("lookup", function(
+                err: Error,
+                address: any,
+                family: any,
+                host: any
+              ) {
+                if (rangeCheck.inRange(address, blacklistedAddresses)) {
+                  res.status(403).send("IP address is not allowed: " + address);
+                  res.end();
+                  proxiedRequest.abort();
+                }
+              });
+            })
+            .on("error", function(err: Error) {
+              console.log(err);
+
+              // Ideally we would return an error to the client, but if headers have already been sent,
+              // attempting to set a status code here will fail. So in that case, we'll just end the response,
+              // for lack of a better option.
+              if (res.headersSent) {
+                res.end();
+              } else {
+                res.status(500).send("Proxy error");
+              }
+            })
+            .on("response", function(response: any) {
+              if (
+                !req.retryWithoutAuth &&
+                response.statusCode === 403 &&
+                proxyAuth[remoteUrl.host]
+              ) {
+                // We automatically added an authentication header to this request (e.g. from proxyauth.json),
+                // but got back a 403, indicating our credentials didn't authorize access to this resource.
+                // Try again without credentials in order to give the user the opportunity to supply
+                // their own.
+                req.retryWithoutAuth = true;
+                return handler(req, res, next);
+              }
+
+              res.status(response.statusCode);
+
+              // Skip caching if we have an error response
+              maxAgeSeconds =
+                response.statusCode >= 400 ? undefined : maxAgeSeconds;
+              res.header(processHeaders(response.headers, maxAgeSeconds));
+
+              response.on("data", function(chunk: any) {
+                res.write(chunk);
+              });
+              response.on("end", function() {
+                res.end();
+              });
+            });
+        } catch (e) {
+          console.error(e.stack);
+          res.status(500).send("Proxy error");
+        }
+
+        return proxiedRequest;
+      });
+    }
+
+    return handler;
+  }
+
+  var router = express.Router();
+  router.get("/*", buildReqHandler("GET"));
+  router.post(
+    "/*",
+    bodyParser.raw({
+      type: function() {
+        return true;
+      },
+      limit: proxyPostSizeLimit
+    }),
+    buildReqHandler("POST")
+  );
+
+  return router;
+};
+
+/**
+ * Filters headers that are not matched by {@link DO_NOT_PROXY_REGEX} out of an object containing headers. This does not
+ * mutate the original list.
+ *
+ * @param headers The headers to filter
+ * @returns {Object} A new object with the filtered headers.
+ */
+function filterHeaders(headers: any) {
+  var result: any = {};
+  // filter out headers that are listed in the regex above
+  Object.keys(headers).forEach(function(name) {
+    if (!DO_NOT_PROXY_REGEX.test(name)) {
+      result[name] = headers[name];
+    }
+  });
+
+  return result;
+}
+
+/**
+ * Filters out headers that shouldn't be proxied, overrides caching so files are retained for {@link DEFAULT_MAX_AGE_SECONDS},
+ * and sets CORS headers to allow all origins
+ *
+ * @param headers The original object of headers. This is not mutated.
+ * @param maxAgeSeconds the amount of time in seconds to cache for. This will override what the original server
+ *          specified because we know better than they do.
+ * @returns {Object} The new headers object.
+ */
+function processHeaders(headers: any, maxAgeSeconds: number) {
+  var result = filterHeaders(headers);
+
+  if (maxAgeSeconds !== undefined) {
+    result["Cache-Control"] = "public,max-age=" + maxAgeSeconds;
+  }
+
+  result["Access-Control-Allow-Origin"] = "*";
+  return result;
+}
